Header Icons changes type define in layout employee edit action completed
