+++ src/components/Header.jsx
@@ -6,18 +6,22 @@ import {
   MenuItem,
   Grid,
   useMediaQuery,
+  Tooltip,
 } from "@mui/material";
 import MenuIcon from "@mui/icons-material/Menu";
 import Image from "next/image";
+import ChatIcon from '@mui/icons-material/Chat';  
 import creworderIcon from "../images/crewordericon.png";
 import ProfileHeader from "./ProfileHeader";
 import SearchBar from "./SearchBar";
+import SettingsIcon from '@mui/icons-material/Settings';
 import { logout } from "@/utils/auth";
+import AccountTreeIcon from '@mui/icons-material/AccountTree';
 import { useRouter } from "next/router";
 import RechargeWallet from "./recharge-wallet/RechargeWallet";
 import BranchSwitcher from "./branchswitcher/BranchSwitcher";
 
-const Header = ({ onMenuClick }) => {
+const Header = ({ onMenuClick, type }) => {
   const [anchorEl, setAnchorEl] = React.useState(null);
   const router = useRouter();
   const isMobile = useMediaQuery("(max-width:600px)");
@@ -39,6 +43,12 @@ const Header = ({ onMenuClick }) => {
     await logout();
     router.push("/login");
     handleMenuClose();
+  };
+
+  const handleItemClick = (path) => {
+    if (path) {
+      router.push(path);
+    }
   };
 
   return (
@@ -89,11 +99,17 @@ const Header = ({ onMenuClick }) => {
           }}
         >
           <Grid container spacing={2} sx={{ alignItems: "center" }}>
-            <Grid item xs={4} sm={4} md={4}>
-              <RechargeWallet /> Wallet Recharge
+            <Grid item sm={6} md={6} sx={{marginLeft: 10}} >
+            <a  style={{ textDecoration: 'none' }}> <Tooltip title="Wallet"><RechargeWallet sx={{marginLeft: 3}} /></Tooltip></a>
+            <a href="/chat" style={{ textDecoration: 'none' }}><Tooltip title="Chat"> <ChatIcon sx={{marginLeft: 3}}/></Tooltip></a>
+
+            {type == "admin" && (<a href="/admin/settings" style={{ textDecoration: 'none' }}> <Tooltip title="Settings"> <SettingsIcon sx={{marginLeft: 3}}/></Tooltip></a>)}
+
+            {type == "superadmin" && ( <a href="/superadmin/settings" style={{ textDecoration: 'none' }}> <Tooltip title="Settings"> <SettingsIcon sx={{marginLeft: 3}}/></Tooltip></a>)}
+            {type == "admin" && (<a href="/admin/branch" style={{ textDecoration: 'none' }}><Tooltip title="Branchs "><AccountTreeIcon sx={{marginLeft: 3}} /></Tooltip></a>)}
             </Grid>
-            <Grid item xs={4} sm={4} md={4}><BranchSwitcher /></Grid>
-            <Grid item xs={4} sm={4} md={4}>
+           
+            <Grid item sm={4} md={4}>
               <Grid
                 container
                 sx={{
