+++ src/components/hr-employees/EditEmployees.jsx
@@ -1,49 +1,49 @@
-import React, { useState, useEffect } from "react";
-import PhotoCamera from "@mui/icons-material/PhotoCamera";
+import React, { useEffect, useState } from "react";
 import CustomTextField from "@/components/CustomTextField";
+import CustomLabel from "../customLabel";
+import CustomCard from "../CustomCard";
+import { useRouter } from "next/router";
+import MainApi from "@/api-manage/MainApi";
+import { getToken } from "@/utils/getToken";
+import useGetAllBranches from "@/api-manage/react-query/useGetAllBranches";
 import useGetAllDepartments from "@/api-manage/react-query/useGetAllDepartments";
 import useGetAllDesignations from "@/api-manage/react-query/useGetAllDesignations";
+import useGetAllUsers from "@/api-manage/react-query/useGetAllUsers";
 import { usePermissions } from "@/contexts/PermissionsContext";
-import CustomLabel from "../customLabel";
 import {
   Typography,
   Button,
   Grid,
-  Card,
   CardContent,
   Divider,
-  Checkbox,
+  MenuItem,
+  Radio,
+  RadioGroup,
+  FormLabel,
   FormControlLabel,
+  Tooltip,
+  List,
+  ListItem,
+  ListItemText,
 } from "@mui/material";
-import CustomCard from "../CustomCard";
-import { useRouter } from "next/router";
-import { MenuItem, Select, InputLabel, FormControl } from "@mui/material";
-import { getToken } from "@/utils/getToken";
-import MainApi from "@/api-manage/MainApi";
-const handleFileChange = (event) => {
-  const file = event.target.files[0];
-  const reader = new FileReader();
-  reader.onload = (e) => {
-    document.getElementById("preview").src = e.target.result;
-  };
-  if (file) {
-    reader.readAsDataURL(file);
-  }
-};
+import { get_username_suggestions } from "@/api-manage/ApiRoutes";
 
 const EditEmployee = () => {
+  const { data: branchData, refetch: branchRefetch } = useGetAllBranches();
+  const { data: departmentData, refetch: departmentRefetch } =
+    useGetAllDepartments();
+  const { data: designationData, refetch: designationRefetch } =
+    useGetAllDesignations();
+  const { data: usersData, refetch: usersRefetch } = useGetAllUsers();
+  const [loading, setLoading] = useState(true);
+  const [employees, setEmployees] = useState(null);
+  const [error, setError] = useState(null);
   const router = useRouter();
   const { id } = router.query;
-  const {permissionsData} = usePermissions();
-  const { data: departmentData, refetch: departmentRefetch } =
-  useGetAllDepartments();
-  const { data: designationData, refetch: designationRefetch } =
-  useGetAllDesignations();
-  const [employees, setEmployees] = useState({
+  const [formData, setFormData] = useState({
     username: "",
     first_name: "",
     last_name: "",
-    profile_image: null,
     email: "",
     date_joined: "",
     profile: {
@@ -51,24 +51,50 @@ const EditEmployee = () => {
       date_of_joining: "",
       date_of_birth: "",
       marital_status: "",
+      professional_email: "",
       contact_no: "",
-      profile_image: null,
-      company: permissionsData?.user?.profile?.company
+      address: "",
+      department: "",
+      designation: "",
+      branch: "",
+      company: "",
+      reporting: "",
+      employment_type: "",
+      enrollment_id: "",
+      reporting: "",
+      employee_id: "",
     },
     role: {
-      role: permissionsData?.role,
+      role: "superadmin",
     },
   });
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
 
-  const handleInputChange = (event) => {
-    const { name, value } = event.target;
-    setEmployees((prevEmployees) => ({
-      ...prevEmployees,
-      [name]: value,
-    }));
-  };
+  useEffect(() => {
+    const fetchEmployee = async () => {
+      if (id) {
+        try {
+          setLoading(true);
+          const token = getToken();
+          const response = await MainApi.get(`/api/users/${id}/`, {
+            headers: {
+              Authorization: `Token ${token}`,
+            },
+          });
+          if (response.status === 200) {
+            // Pre-fill form with fetched data
+            setFormData(response.data);
+          } else {
+            console.error("Failed to fetch user data");
+          }
+        } catch (error) {
+          console.error("Error fetching user data:", error);
+        } finally {
+          setLoading(false);
+        }
+      }
+    };
+    fetchEmployee();
+  }, [id]);
 
   const handleSubmit = async (event) => {
     event.preventDefault();
@@ -79,30 +105,45 @@ const EditEmployee = () => {
         throw new Error("No authentication token found.");
       }
 
-      const formData = new FormData();
-      for (const key in employees) {
-        formData.append(key, employees[key]);
-      }
+      const compareFields = (formData, originalData) => {
+        const updatedFields = {};
 
-      console.log("Employee Form Data", formData);
+        for (const key in formData) {
+          if (typeof formData[key] === "object" && formData[key] !== null) {
+            const nestedUpdatedFields = compareFields(
+              formData[key],
+              originalData[key]
+            );
+            if (Object.keys(nestedUpdatedFields).length > 0) {
+              updatedFields[key] = nestedUpdatedFields;
+            }
+          } else if (formData[key] !== originalData[key]) {
+            updatedFields[key] = formData[key];
+          }
+        }
 
-      const response = await MainApi.put(`/api/users/${id}/`, formData, {
+        return updatedFields;
+      };
+
+      const updatedFields = compareFields(formData, employees);
+      console.log("Updated Fields", updatedFields);
+      const response = await MainApi.patch(`/api/users/${id}/`, updatedFields, {
         headers: {
           Authorization: `Token ${token}`,
-          "Content-Type": "multipart/form-data",
+          "Content-Type": "application/json",
         },
       });
 
       if (response.status === 200) {
-        console.log("Employees updated successfully");
-        router.push("/admin/employees");
+        console.log("Employee updated successfully");
+         router.push("/hr/employees");
       } else {
-        console.error("Failed to update the Employees");
-        setError("Failed to update the Employees");
+        console.error("Failed to update the employee");
+        setError("Failed to update the employee");
       }
     } catch (error) {
-      console.error("An error occurred while updating the Employees:", error);
-      setError("An error occurred while updating the Employees");
+      console.error("An error occurred while updating the employee:", error);
+      setError("An error occurred while updating the employee");
     }
   };
 
@@ -123,205 +164,88 @@ const EditEmployee = () => {
           });
 
           if (response.status === 200) {
-            setEmployees(response.data);
+            const employeeData = response.data;
+            setEmployees(employeeData); // Initialize employees with fetched data
+            setFormData(employeeData);
           } else {
-            console.error("Failed to fetch the employee");
-            setError("Failed to fetch the employee");
+            console.error("Failed to fetch the Agent");
+            setError("Failed to fetch the Agent");
           }
         } catch (error) {
-          console.error(
-            "An error occurred while fetching the employee:",
-            error
-          );
-          setError("An error occurred while fetching the employee");
+          console.error("An error occurred while fetching the Agent:", error);
+          setError("An error occurred while fetching the Agent");
         } finally {
           setLoading(false);
         }
       };
-
       fetchEmployee();
     }
   }, [id]);
 
-  const formatDate = (dateString) => {
-    var date = new Date(dateString);
-    date = date.toLocaleDateString();
-    const parts = date.split("/");
-  
-  // Check if we have exactly three parts (month, day, year)
-  if (parts.length !== 3) {
-    console.error("Invalid date format. Expected format: 'M/D/YYYY'");
-    return null;
-  }
+  const [errors, setErrors] = useState({});
+  const [suggestions, setSuggestions] = useState([]);
+  const handleRadioChange = (event) => {
+    setFormData({ ...formData, loginAllowed: event.target.value });
+  };
 
-  // Extract month, day, and year from the split parts
-  const [month, day, year] = parts;
+  const handleSuggestionClick = (suggestion) => {
+    setFormData((prevData) => ({ ...prevData, username: suggestion }));
+    setSuggestions([]);
+  };
 
-  // Check if month, day, and year are not undefined
-  if (!month || !day || !year) {
-    console.error("Invalid date format. Month, day, or year is missing.");
-    return null;
-  }
-
-  // Pad month and day with leading zeros if necessary
-  const paddedMonth = month.padStart(2, '0');
-  const paddedDay = day.padStart(2, '0');
-
-  // Format the date in "YYYY-MM-DD"
-  return `${year}-${paddedMonth}-${paddedDay}`;
+  const handleChange = (e) => {
+    setFormData({ ...formData, [e.target.name]: e.target.value });
   };
 
   return (
     <Grid container sx={{ padding: 3 }}>
       <Grid item xs={12}>
         <CustomCard>
-          <Grid item>
-            <Typography
-              sx={{ mt: 2, ml: 2, fontSize: "16px", fontWeight: "600" }}
-            >
-              Update Employee .
-            </Typography>
-          </Grid>
-          <Divider sx={{ my: 2 }} />
-
           <CardContent>
-            <Grid item>
-              <Typography sx={{ fontSize: "16px", fontWeight: "600" }}>
-                Personal Information.
-              </Typography>
-            </Grid>
+            <Typography sx={{ fontSize: "18px", fontWeight: "600" }}>
+              Edit Employee
+            </Typography>
             <Divider sx={{ my: 2 }} />
-            <Grid
-              item
-              sx={{
-                display: "flex",
-                flexDirection: { xs: "column", sm: "row" },
-                gap: 2,
-                marginTop: 2,
-              }}
-            >
-              <Grid item sx={{ flex: 1 }}>
-                <CustomLabel htmlFor="username" required>
-                  Username: :
-                </CustomLabel>
-                <CustomTextField
-                  id="username"
-                  name="username"
-                  type="text"
-                  placeholder="Enter UserName"
-                  required
-                  fullWidth
-                  value={employees.username}
-                  onChange={handleInputChange}
-                />
+            <Grid container spacing={2}>
+              <Grid item xs={12}>
+                <Typography sx={{ fontSize: "16px", fontWeight: "600" }}>
+                  Personal Info.
+                </Typography>
               </Grid>
 
-              
-              <Grid item sx={{ flex: 1 }}>
-                <CustomLabel htmlFor="password" required>
-                  Password: :
-                </CustomLabel>
-                <CustomTextField
-                  id="password"
-                  name="password"
-                  type="password"
-                  required
-                  fullWidth
-                  value={employees.password}
-                  onChange={handleInputChange}
-                />
-              </Grid>
-        
-
-
-              <Grid item sx={{ flex: 1 }}>
+              <Grid item xs={12} sm={4}>
                 <CustomLabel htmlFor="first_name" required>
-                  First Name
+                  First Name:
                 </CustomLabel>
                 <CustomTextField
                   id="first_name"
                   name="first_name"
-                  placeholder="First Name"
+                  placeholder="e.g. firstname"
                   type="text"
-                  value={employees.first_name}
-                  onChange={handleInputChange}
-                  required
                   fullWidth
+                  value={formData.first_name || ""}
+                  onChange={handleChange}
+                  error={!!errors.first_name}
+                  helperText={errors.first_name}
                 />
               </Grid>
-
-              <Grid item sx={{ flex: 1 }}>
+              <Grid item xs={12} sm={4}>
                 <CustomLabel htmlFor="last_name" required>
-                  Last Name
+                  Last Name:
                 </CustomLabel>
                 <CustomTextField
                   id="last_name"
                   name="last_name"
-                  placeholder="Last Name"
                   type="text"
-                  value={employees.last_name}
-                  onChange={handleInputChange}
-                  required
+                  placeholder="e.g. lastname"
                   fullWidth
+                  value={formData.last_name || ""}
+                  onChange={handleChange}
+                  error={!!errors.last_name}
+                  helperText={errors.last_name}
                 />
               </Grid>
-            </Grid>
-
-            <Grid
-              item
-              sx={{
-                display: "flex",
-                flexDirection: { xs: "column", sm: "row" },
-                gap: 2,
-                marginTop: 2,
-              }}
-            >
               <Grid item xs={12} sm={4}>
-                <CustomLabel htmlFor="category" required>
-                  Gender
-                </CustomLabel>
-                <Select
-                  labelId="gender"
-                  id="gender"
-                  name="gender"
-                  value={employees.profile.gender}
-                  onChange={handleInputChange}
-                  displayEmpty
-                  sx={{ fontFamily: "Poppins, sans-serif", height: "40px" }}
-                  fullWidth
-                >
-                  <MenuItem value="" disabled>
-                    Select Gender
-                  </MenuItem>
-                  <MenuItem value="m">Male</MenuItem>
-                  <MenuItem value="f">Female</MenuItem>
-                  <MenuItem value="t">Other</MenuItem>
-                </Select>
-              </Grid>
-
-              <Grid item xs={12} sm={4}>
-                <CustomLabel htmlFor="category" required>
-                  Marital Status
-                </CustomLabel>
-                <Select
-                  labelId="gender"
-                  id="gender"
-                  name="gender"
-                  value={employees.profile.marital_status}
-                  onChange={handleInputChange}
-                  displayEmpty
-                  sx={{ fontFamily: "Poppins, sans-serif", height: "40px" }}
-                  fullWidth
-                >
-                  <MenuItem value="" disabled>
-                    Select Marital Status
-                  </MenuItem>
-                  <MenuItem value="married">Married</MenuItem>
-                  <MenuItem value="unmarried">UnMarried</MenuItem>
-                </Select>
-              </Grid>
-
-              <Grid item sx={{ flex: 1 }}>
                 <CustomLabel htmlFor="dob" required>
                   Date of Birth:
                 </CustomLabel>
@@ -329,171 +253,114 @@ const EditEmployee = () => {
                   id="dob"
                   name="dob"
                   type="date"
-                  value={employees.profile.date_of_birth}
-                  onChange={handleInputChange}
-                  required
                   fullWidth
+                  value={formData?.profile?.date_of_birth}
+                  onChange={handleChange}
+                  error={!!errors.dob}
+                  helperText={errors.dob}
                 />
               </Grid>
             </Grid>
 
-            <Grid
-              item
-              sx={{
-                display: "flex",
-                flexDirection: { xs: "column", sm: "row" },
-                gap: 2,
-                marginTop: 2,
-              }}
-            >
+            <Grid container spacing={2} mt={1}>
               <Grid item xs={12} sm={4}>
-                <CustomLabel htmlFor="phone" required>
-                  Phone Number
+                <CustomLabel htmlFor="gender" required>
+                  Select Gender
                 </CustomLabel>
                 <CustomTextField
-                  id="phone"
-                  name="phone"
-                  type="text"
-                  placeholder="(+91)"
-                  value={employees.profile.contact_no}
-                  onChange={handleInputChange}
-                  required
+                  id="gender"
+                  name="gender"
+                  select
                   fullWidth
+                  value={formData?.profile?.gender}
+                  onChange={handleChange}
+                  error={!!errors.gender}
+                  helperText={errors.gender}
+                >
+                  <MenuItem value="Select Gender" disabled>
+                    Select Gender
+                  </MenuItem>
+                  <MenuItem value="m">Male</MenuItem>
+                  <MenuItem value="f">Female</MenuItem>
+                  <MenuItem value="o">Other</MenuItem>
+                </CustomTextField>
+              </Grid>
+
+              <Grid item xs={12} sm={4}>
+                <CustomLabel htmlFor="maritalStatus" required>
+                  Marital Status
+                </CustomLabel>
+                <CustomTextField
+                  id="maritalStatus"
+                  name="maritalStatus"
+                  select
+                  fullWidth
+                  value={formData?.profile?.marital_status}
+                  onChange={handleChange}
+                  error={!!errors.maritalStatus}
+                  helperText={errors.maritalStatus}
+                >
+                  <MenuItem value="Select Marital Status" disabled>
+                    Select Marital Status
+                  </MenuItem>
+                  <MenuItem value="married">Married</MenuItem>
+                  <MenuItem value="unmarried">Unmarried</MenuItem>
+                </CustomTextField>
+              </Grid>
+
+              <Grid item xs={12} sm={4}>
+                <CustomLabel htmlFor="contact_number" required>
+                  Contact Number
+                </CustomLabel>
+                <CustomTextField
+                  id="contact_number"
+                  name="contact_number"
+                  type="tel"
+                  placeholder="e.g. 9999999999"
+                  fullWidth
+                  value={formData?.profile?.contact_no}
+                  onChange={handleChange}
+                  error={!!errors.contact_number}
+                  helperText={errors.contact_number}
                 />
               </Grid>
+            </Grid>
+
+            <Grid container spacing={2} mt={2}>
               <Grid item xs={12} sm={4}>
                 <CustomLabel htmlFor="email" required>
-                  Email
+                  Personal Email:
                 </CustomLabel>
                 <CustomTextField
                   id="email"
                   name="email"
                   type="email"
-                  placeholder="e.g. test@creworder.com"
-                  value={employees.email}
-                  onChange={handleInputChange}
-                  required
+                  placeholder="e.g. email@example.com"
                   fullWidth
-                  sx={{ height: 40 }}
+                  value={formData.email || ""}
+                  onChange={handleChange}
+                  error={!!errors.email}
+                  helperText={errors.email}
                 />
               </Grid>
 
               <Grid item xs={12} sm={4}>
-                <CustomLabel htmlFor="profile_images" required>
-                  Upload Profile
-                </CustomLabel>
-
-                <img
-                  id="preview"
-                  src="https://static.vecteezy.com/system/resources/thumbnails/021/353/308/small_2x/user-icon-for-website-and-mobile-apps-png.png"
-                  alt="Preview"
-                  width="35%"
-                  style={{ objectFit: "contain" }}
-                />
-                <input
-                  type="file"
-                  id="profile_images"
-                  onChange={handleFileChange}
-                  style={{ marginTop: "8px", display: "block", width: "100%" }}
-                />
-              </Grid>
-            </Grid>
-
-            <Grid
-              item
-              sx={{
-                display: "flex",
-                flexDirection: { xs: "column", sm: "row" },
-                gap: 2,
-                marginTop: 2,
-              }}
-            >
-              <Grid item xs={12} sm={12}>
-                <CustomLabel htmlFor="address" required>
-                  Address:
+                <CustomLabel htmlFor="professionalemail" required>
+                  Professional Email:
                 </CustomLabel>
                 <CustomTextField
-                  id="address"
-                  name="address"
-                  type="text"
-                  placeholder="Address"
-                  value={employees.profile.address}
-                  onChange={handleInputChange}
-                  required
+                  id="professionalemail"
+                  name="professionalemail"
+                  type="email"
+                  placeholder="e.g. email@example.com"
                   fullWidth
-                  multiline
-                  rows={4} // This makes it a textarea
+                  value={formData?.profile?.professional_email}
+                  onChange={handleChange}
+                  error={!!errors.professionalemail}
+                  helperText={errors.professionalemail}
                 />
               </Grid>
-            </Grid>
-            <Grid container>
-              {/* Other components/content above */}
-              <Grid item xs={12}>
-                <Divider sx={{ marginBottom: 2 }} />
-              </Grid>
-              <Grid item>
-                <Typography sx={{ fontSize: "16px", fontWeight: "600" }}>
-                  Office Information.
-                </Typography>
-              </Grid>
-              {/* Other components/content below */}
-            </Grid>
-            <Divider sx={{ my: 2 }} />
-            <Grid
-              item
-              sx={{
-                display: "flex",
-                flexDirection: { xs: "column", sm: "row" },
-                gap: 2,
-                marginTop: 2,
-              }}
-            >
-              <Grid item sx={{ flex: 1 }}>
-                <CustomLabel htmlFor="designation" required>
-                  Department:
-                </CustomLabel>
-                <Select
-                  labelId="gender"
-                  id="gender"
-                  name="gender"
-                  value={employees.profile.department}
-                  onChange={handleInputChange}
-                  displayEmpty
-                  sx={{ fontFamily: "Poppins, sans-serif", height: "40px" }}
-                  fullWidth
-                >
-                  <MenuItem value="" disabled>
-                    Select Department
-                  </MenuItem>
-                  {departmentData?.results.map((row, index) => (
-                      <MenuItem value={row?.id}>{row.name}</MenuItem>
-                    ))}
-                </Select>
-              </Grid>
-
-              <Grid item sx={{ flex: 1 }}>
-                <CustomLabel htmlFor="designation" required>
-                  Designation:
-                </CustomLabel>
-                <Select
-                  labelId="gender"
-                  id="gender"
-                  name="gender"
-                  value={employees.profile.designation}
-                  onChange={handleInputChange}
-                  displayEmpty
-                  sx={{ fontFamily: "Poppins, sans-serif", height: "40px" }}
-                  fullWidth
-                >
-                  <MenuItem value="" disabled>
-                    Select Designation
-                  </MenuItem>
-                  {designationData?.results.map((row, index) => (
-                      <MenuItem value={row?.id}>{row.name}</MenuItem>
-                    ))}
-                </Select>
-              </Grid>
-              <Grid item sx={{ flex: 1 }}>
+              <Grid item xs={12} sm={4}>
                 <CustomLabel htmlFor="doj" required>
                   Date of Joining:
                 </CustomLabel>
@@ -501,38 +368,248 @@ const EditEmployee = () => {
                   id="doj"
                   name="doj"
                   type="date"
-                  // value="1984-08-09"
-                  value={formatDate(employees.profile.date_of_joining)}
-                  onChange={handleInputChange}
-                  required
                   fullWidth
+                  value={formData?.profile?.date_of_joining}
+                  onChange={handleChange}
+                  error={!!errors.doj}
+                  helperText={errors.doj}
                 />
               </Grid>
             </Grid>
 
-            <Grid
-              item
-              sx={{
-                marginTop: 3,
-                display: "flex",
-                justifyContent: "flex-end",
-              }}
-            >
-              <Button
-                sx={{
-                  padding: "8px 16px",
-                  fontSize: "14px",
-                  fontWeight: "bold",
-                  backgroundColor: "#405189",
-                  color: "white",
-                  "&:hover": {
-                    backgroundColor: "#334a6c",
-                  },
-                }}
-                onClick={handleSubmit}
-              >
-                Submit
-              </Button>
+            <Grid container spacing={2} mt={2}>
+              <Grid item xs={12} sm={8}>
+                <CustomLabel htmlFor="address" required>
+                  Address:
+                </CustomLabel>
+                <CustomTextField
+                  id="address"
+                  name="address"
+                  type="text"
+                  placeholder="e.g. Address"
+                  fullWidth
+                  multiline
+                  rows={2}
+                  value={formData?.profile?.address}
+                  onChange={handleChange}
+                  error={!!errors.address}
+                  helperText={errors.address}
+                />
+              </Grid>
+
+              <Grid item xs={12} sm={4}>
+                <CustomLabel htmlFor="department" required>
+                  Department
+                </CustomLabel>
+                <CustomTextField
+                  id="department"
+                  name="department"
+                  select
+                  fullWidth
+                  value={formData?.profile?.department}
+                  onChange={handleChange}
+                  error={!!errors.department}
+                  helperText={errors.department}
+                >
+                  <MenuItem value="Select Department" disabled>
+                    Select Department
+                  </MenuItem>
+                  {departmentData?.results.map((row, index) => (
+                    <MenuItem value={row?.id}>{row.name}</MenuItem>
+                  ))}
+                </CustomTextField>
+              </Grid>
+            </Grid>
+
+            <Grid container spacing={2} mt={2}>
+              <Grid item xs={12} sm={6}>
+                <CustomLabel htmlFor="designation" required>
+                  Designation
+                </CustomLabel>
+                <CustomTextField
+                  id="designation"
+                  name="designation"
+                  select
+                  fullWidth
+                  value={formData?.profile?.designation}
+                  onChange={handleChange}
+                  error={!!errors.designation}
+                  helperText={errors.designation}
+                >
+                  <MenuItem value="Select Designation" disabled>
+                    Select Designation
+                  </MenuItem>
+                  {designationData?.results.map((row, index) => (
+                    <MenuItem value={row?.id}>{row.name}</MenuItem>
+                  ))}
+                </CustomTextField>
+              </Grid>
+
+              <Grid item xs={12} sm={6}>
+                <CustomLabel htmlFor="reportingto" required>
+                  Reporting To
+                </CustomLabel>
+                <CustomTextField
+                  id="reportingto"
+                  name="reportingto"
+                  select
+                  fullWidth
+                  value={formData?.profile?.reporting}
+                  onChange={handleChange}
+                  error={!!errors.reportingto}
+                  helperText={errors.reportingto}
+                >
+                  <MenuItem value="Select Reporting To" disabled>
+                    Select Reporting To
+                  </MenuItem>
+                  {usersData?.results.map((row, index) => (
+                    <MenuItem value={row?.id}>{row.username}</MenuItem>
+                  ))}
+                </CustomTextField>
+              </Grid>
+            </Grid>
+
+            <Grid container spacing={2} mt={2}>
+              <Grid item xs={12} sm={6}>
+                <CustomLabel htmlFor="branch" required>
+                  Branch
+                </CustomLabel>
+                <CustomTextField
+                  id="branch"
+                  name="branch"
+                  select
+                  fullWidth
+                  value={formData?.profile?.branch}
+                  onChange={handleChange}
+                  error={!!errors.branch}
+                  helperText={errors.branch}
+                >
+                  <MenuItem value="Select Branch" disabled>
+                    Select Branch
+                  </MenuItem>
+                  {branchData?.results.map((row, index) => (
+                    <MenuItem value={row?.id}>{row.name}</MenuItem>
+                  ))}
+                </CustomTextField>
+              </Grid>
+
+              <Grid item xs={12} sm={6}>
+                <CustomLabel htmlFor="employeeType" required>
+                  Employee Type
+                </CustomLabel>
+                <CustomTextField
+                  id="employeeType"
+                  name="employeeType"
+                  select
+                  fullWidth
+                  value={formData?.profile?.employment_type}
+                  onChange={handleChange}
+                  error={!!errors.employeeType}
+                  helperText={errors.employeeType}
+                >
+                  <MenuItem value="Employee Type" disabled>
+                    Employee Type
+                  </MenuItem>
+                  <MenuItem value="full">Full-time</MenuItem>
+                  <MenuItem value="part">Part-time</MenuItem>
+                  <MenuItem value="trainee">Trainee</MenuItem>
+                  <MenuItem value="intern">Internship</MenuItem>
+                </CustomTextField>
+              </Grid>
+            </Grid>
+
+            <Typography sx={{ fontSize: "18px", fontWeight: "600" }}>
+              User Details
+            </Typography>
+            <Divider sx={{ my: 2 }} />
+
+            <Grid container spacing={2} mt={2}>
+              <Grid item xs={12} md={4}>
+                <CustomLabel htmlFor="username" required>
+                  Username:
+                </CustomLabel>
+
+                <CustomTextField
+                  id="username"
+                  name="username"
+                  type="text"
+                  placeholder="Enter Username"
+                  value={formData.username}
+                  onChange={handleChange}
+                  error={!!errors.username}
+                  helperText={errors.username}
+                  fullWidth
+                  inputProps={{ readOnly: true }}
+                />
+              </Grid>
+              <Grid item xs={12} md={4}>
+                {suggestions.length > 0 && (
+                  <List>
+                    {suggestions.map((suggestion, index) => (
+                      <ListItem
+                        key={index}
+                        button
+                        onClick={() => handleSuggestionClick(suggestion)}
+                      >
+                        <ListItemText primary={suggestion} />
+                      </ListItem>
+                    ))}
+                  </List>
+                )}
+              </Grid>
+
+              <Grid item xs={12} md={4}>
+                <CustomLabel htmlFor="password" required>
+                  Password:
+                </CustomLabel>
+                <CustomTextField
+                  id="password"
+                  name="password"
+                  type="text"
+                  placeholder="Enter password"
+                  value={formData.password}
+                  onChange={handleChange}
+                  error={!!errors.password}
+                  helperText={errors.password}
+                  fullWidth
+                />
+              </Grid>
+              <Grid item xs={4}>
+                <FormLabel component="legend">
+                  <b>Login Allowed:</b>
+                </FormLabel>
+                <RadioGroup
+                  aria-label="loginAllowed"
+                  name="loginAllowed"
+                  value={String(formData?.profile?.login_allowed)} // Convert boolean to string
+                  onChange={(e) => handleRadioChange(e, "login_allowed")} // Modify onChange
+                  row
+                >
+                  <FormControlLabel
+                    value="true"
+                    control={<Radio />}
+                    label="Yes"
+                  />
+                  <FormControlLabel
+                    value="false"
+                    control={<Radio />}
+                    label="No"
+                  />
+                </RadioGroup>
+              </Grid>
+            </Grid>
+
+            <Grid container spacing={2} mt={2}>
+              <Grid item xs={4}>
+                <Button
+                  variant="contained"
+                  color="primary"
+                  fullWidth
+                  onClick={handleSubmit}
+                >
+                  Update Agent
+                </Button>
+              </Grid>
             </Grid>
           </CardContent>
         </CustomCard>
@@ -542,3 +619,4 @@ const EditEmployee = () => {
 };
 
 export default EditEmployee;
+
