--- src/pages/admin/agent/createagent.js
@@ -1,12 +1,0 @@
-import CreateAgent from '@/components/agent/CreateAgent'
-import React from 'react'
-import Layout from '@/components/Layout'
-const createagent = () => {
-  return (
-    <Layout type="admin">
-        <CreateAgent />
-    </Layout>
-  )
-}
-
-export default createagent;
