--- src/pages/admin/agent/editagent.js
@@ -1,12 +1,0 @@
-import EditAgent from '@/components/agent/EditAgent'
-import React from 'react'
-import Layout from '@/components/Layout'
-const editagent = () => {
-  return (
-    <Layout type="admin">
-        <EditAgent />
-    </Layout>
-  )
-}
-
-export default editagent;
