--- src/pages/admin/agent/index.js
@@ -1,12 +1,0 @@
-import AgentList from '@/components/agent/AgentList'
-import React from 'react'
-import Layout from '@/components/Layout'
-const index = () => {
-  return (
-    <Layout type="admin">
-    <AgentList />
-    </Layout>
-  )
-}
-
-export default index
